pyproject.toml
[project]
name = "lastwar-first-lady-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyautogui>=0.9.54",
    "pyyaml>=6.0.2",
]
